Adding yamnet support

+++ setup/scripts/ss-setup.sh
@@ -40,7 +40,8 @@ fi
 sudo cp ss-apps/setup/systemd/*.service /lib/systemd/system/
 sudo cp ss-apps/setup/scripts/*.sh /usr/local/bin
 sudo systemctl daemon-reload
-sudo systemctl enable ss-inf
+#sudo systemctl enable ss-inf
+sudo systemctl enable ss-yamnet-inf
 sudo systemctl enable ss-led
 sudo systemctl enable ss-heartbeat
 #sudo systemctl enable ss-gui
@@ -86,18 +87,22 @@ declare -a models=("hio5-nobaby_sigmoid_9_5_64.tflite" \
                    "hio5-nochild-classical_sigmoid_9_5_96.tflite"  \
                    "hio5-nobaby-classical_softmax_9_5_96.tflite" \
                    "hio5-nobaby-classical_sigmoid_9_5_96.tflite")
-for i in "${models[@]}"
-do
-	sudo curl -XGET -o ${ss_dir}/"${i}" "https://www.googleapis.com/storage/v1/b/ss-service-models/o/${i}?alt=media"
-done
-
-if [ -e "${ss_dir}/soundscene.tflite" ]; then
-	sudo rm  ${ss_dir}/soundscene.tflite
+                   
+read -p "Would you like to install SS Models? (y/N)" -n 1 -r -s
+if [[ $REPLY =~ ^[Yy]$ ]]; then
+	for i in "${models[@]}"
+	do
+		sudo curl -XGET -o ${ss_dir}/"${i}" "https://www.googleapis.com/storage/v1/b/ss-service-models/o/${i}?alt=media"
+	done
+	
+	
+	if [ -e "${ss_dir}/soundscene.tflite" ]; then
+		sudo rm  ${ss_dir}/soundscene.tflite
+	fi
+	sudo ln -s  ${ss_dir}/${models[0]}  ${ss_dir}/soundscene.tflite
+	sudo curl -XGET -o ${ss_dir}/vggish.tflite "https://www.googleapis.com/storage/v1/b/ss-service-models/o/vggish.tflite?alt=media"
+	sudo curl -XGET -o ${audioset_dir}/class_labels_indices.csv "http://storage.googleapis.com/us_audioset/youtube_corpus/v1/csv/class_labels_indices.csv"
 fi
-sudo ln -s  ${ss_dir}/${models[0]}  ${ss_dir}/soundscene.tflite
-sudo curl -XGET -o ${ss_dir}/vggish.tflite "https://www.googleapis.com/storage/v1/b/ss-service-models/o/vggish.tflite?alt=media"
-sudo curl -XGET -o ${audioset_dir}/class_labels_indices.csv "http://storage.googleapis.com/us_audioset/youtube_corpus/v1/csv/class_labels_indices.csv"
-
 #TODO - Install from clone above?
 #install ss
 declare -a pkgs=("pi-core" "pi-apps/leds" "pi-apps/br" "pi-apps/inf" "pi-apps/status" "pi-apps/debug" "pi-web/inf_gui" "pi-svc/audio_playback")
